Fix issue 336. Reordering events
Timeline template has now a link to the dedicated page for events reordering.
Events without a date (or text date) can be moved up and down, dated events
are always in the same order (except events with exactly same date: they can
be swapped only with each other). After applying changes by user new order
is sent to the server where new order is being validated and commited.

Page for updating a person is slightly changed. Event positing in the list
(a.k.a. weight) is exposed to the form. Form hadling code uses this weight to
reconstitute order of the events and to distinguish old events from new ones.
Before that all new events were automatically sorted. A warning about changed
event order is removed.

Also type alias for cdate was removed.