fix incorrect request on image update
v7 plugin used to give an incorrect request if get_image was not an empty string. (why?)
This bug did not appears before because image path is not written to the db
(if it is the default path) and thus get_image often gives an empty string.
Using the Image module instead of get_image made it always buggy