Update lexicon.txt to cause proper dates in Russian
The tables on (day month year) and on (weakday day month year) in hd/lang/lexicon.txt was missing entries for Russian causing display in the default language.
Entries are added. An empty entry for the first and a в for the second, (even if it would be better with во on Tuesdays. 

Solves issue #1471