fix very hard to find %? variable bug!!
Clean some conditions and switch the 4 table_row (function of 11 variables!) to
“long form”; removes a probable parsing bug with “parentheses form”. The traces
are left before the %apply calls in comments because it was very hard (!!) and
it could be usefull for the following. The hyperlink to go back to parents is
now added on every descendants and not only on the elder of each siblings. Add
eight more pixels on margin-left of the container (pl-4 instead of default 3)
to support home on left.