particles management in books of surnames with improved anchors
* Particles are now managed in the book of surname and particles.txt has been
completed.
* Particles present in that file are written at the end of entries in parentheses
but the normal form of a surname is restored in the input itself before renaming.
* The anchor system to go back to the edited entry is fixed when renaming to a
first letters with accentuated characters. If the first letter is changed, two
links are given to fallback either to the new modified entry or to the previous
letter (to continue editing there).
* The anchor system now uses the first two letters of the entries for s=.. fallback.
This allows to bypass the number of entries limitation more often and to fallback
more frequently on the right modified entry.
* Renaming of unknown surnames and first_names (sn=? or fn=?) now don’t rename
unknowns  individuals "? ?" at all.