add Lithuanian and Arabic support
Lithuanian was missing on index and for nearly all other languages.
It was also breaking the links to select Turkish and Chinese in the language dropdown menu.
Arabic had already 38 translations, but LTR for he/ar is badly supported.
Removed 50 white stars and 7 red stripes (en-us had one entry for “color“).