add %apply;url_set based on url_set_aux with forced slash as separator
This second longer form allows:
* To write complex template conditions code for query parameters values.
* Those query parameters stay at their place in the URLs.
* The slash "/" separator handle parameters that include underscores "_".

Like all %apply;function it works with …(param1/param2/…,value1/value2/…)
or when using conditions prefer the form …%with;param1/…%and;value1/…%end;.

The short %url_set.evar.str; function stays unchanged with "_" as separarator.
You obviously can't use it if any of the parameter has an underscore!