new page “conf” for wizards with gwd launch arguments and gwf parameters
Wizards can access a new template file conf.txt that provide information on
the user and languages, a list of the launch arguments of the gwd server and
followed by all base configuration parameters (duplicates are shown in red).
Added new translations for the title and all subtitles of that new page.