new functions to improve shortcuts from forms to book usability
— The ctrl-click shortcut from inputs in forms goes directly to the edited entry
in books, facilitating mass renaming or fixing erroneous entries with around suggestions.
It handles subplaces specificities for the book of places.
— Add a ‘clear button’ (X) to all inputs to facilitate autocompletion: it allows users
to quickly empty the inputs and to see all options offered by the cached datalist.
It supports a new s2 parameter to change the book input from URLs.
— A generic function to put the cursor in inputs with an autofocus parameter at the end of the input.
— A last function highlights greek and cyrilic initial in books using regex.min.js.
This library could be later usefull in books to better identify script mixup, for example inputs that
are in latin and greek and that should not exist.