use precompiled Ocaml, add update_nldb and fix home for Colab
Using precompiled Ocaml 4.13.1 on Colab saves about 5 min of build.
The welcome page with an empty query “…/mybase?” is broken on Colab because it
is adding ?auth=0 automaticaly and this parameter is not interpreted by gwd.
To avoid failure we target “?m=H&v=welcome” at start and also hack the first
hyperlink button of home template with a perl regexp.