Do not expand Makefile variables in double-quoted shell strings
Currently, an attacker could execute arbitrary code on user computers if
he can push a commit message on master. This is not a huge security
issue because the attacker needs to push its commit on master but still
we do not want to execute shell commands here.

Consider the following Makefile example:
```
VAR:=$(shell echo '`ls -l`')

foo:
	@printf "$(VAR)"

.PHONY: foo
```
We expect that `make foo` will print:
`ls -l`

Actually, it will execute the ls command and print the result on stdout!
Indeed, $(VAR) will be expand to `ls -l` and the command:
printf "`ls -l`"
will be executed by bash.

Now, in the current version of Makefile, we read the last commit message
and print it. In particular, if you type markdown in your commit, as I
am used to do, `make distrib` will execute the commands from it!

I replace some of the double quotes by simple quote.