Use `List.rev_map` in `first_name_print_list` and `search_first_name_print`
Searching `Pierre` as a first name in the Roglo database throws a Stack
overflow exception.

The `List.map` and `List.fold_right` functions used in `first_name_print_list`
and `search_first_name_print` are not tail recursive.

In OCaml 4, the usual workaround is to use a combinaison of
`List.rev_map` and `List.rev`. Both of them are tail recursive, but the
code is bit slower (one needs to iterate twice on the list).

In `first_name_print_list`, we do not need to preserve the order
before the end of the function because the list is sorted before calling
`print_alphab_list`.

In `search_first_name_print`, we do not need to preserve the order
because we call `first_name_print_list` at the end. To ensure robustmess,
I use a combinaison of `List.rev_map` and `List.fold_left`, instead of `List.map`
and `List.fold_right`. The produced list is the same, it is tail recursive and
we preserve the same performance.