Add proper hashers for types of Def and Adef
GeneWeb needs to hash these values for digest challenges in some forms.
The Marshal module of the stdlib is not compatible with Ancient and
cannot be used to convert Ancient value into strings before hashing
them.

The module `Hasher` provides hash functions for most of the types
defined in `Adef` and `Def`. These hash functions can be combine without
bulting long strings as the previous approach did.

These hashers use SHA256 algorithm, which is slower than MD5 but safer.
MD5 should never be used as a cryptographic hash function.