Properly clear signal alarm in gwd
The gwd webserver uses the Unix signal alarm to set a timelimit per
request. This feature does not work properly in no-fork mode because the
signal is not reset after treating the request and the signal handler
tries to output to a closed socket. This PR refactors the way we manage
the signal to ensure that we always unset the signal after treating the
request.