Properly close the database
The current API for opening and closing the database is quite fragile.
Developers must manually close it, and it was not done consistently in
many places.

In particular, the database was not closed after loading the database
with Ancient, leading to a warning message in present of several
databases. A one-line PR could fix this issue but I believe we should
implement the a proper solution rather than applying another hot fix.

This PR replaces the `openbase` by a `with_database` and removes the
`close` function. To use a database named `bname`, you now write:
```ocaml
with_database bname (fun base -> ...)
```
The last argument contains your code that uses `base` and `with_database`
ensures that all channels are closed and performs some cleanup operations
before closing them.

If you want to load the base in memory, you must invoke `load_database
bname` before calling `with_database`. In this case, the base is
persistant and `with_database` won't unload it after the continuation is
executed.

Alas, the old API was used at many places, making this PR is quite long but
completely trivial.