Add simple worker pool in Gwd
The gwd server is quite slow when loading pages with multiple images.
After profiling, it appears that the server spends a lot of time
creating forks for each request. This PR adds a simple worker pool to
remove this bottleneck.

Before this commit, the `Louis XIV` request took about 15 seconds on
`test.roglo.eu`. After this change, it took about 5 seconds. Notice that the
first request takes longer because of the computation of height of tree of
descendances. If you want to test this PR, you need to disable the cache
of your browser and reload the page. If your cache is not properly
disabled, your browser won't request images again.

Besides, the option `-max_clients` does not behave as it should. It was
a limit on the number of forks performed by the server, which is not the
number of the clients.

After this PR, `-max_clients` is no-op and displays a warning message.
The option is deprecated and should be removed in a future release.

I add a new option `-n_workers` which gives the number of workers created
while launching the server.

I add another option to control the maximum size of the pending
requests. If the pending queue is full, new request are simply ignored.