Privatize the environment type of Templ
This commit creates a new abstraction barrier for the environments of
Templ.

The `enclosure` variable in `Perso` seems to be unused and its code
cannot be translate easily without assuming that the environment of Templ is an
associated list. This commit removes this code and the delimiters
`#loop`, which were used only by `enclosure`. The purpose of `enclosure`
is unclear and the git history is not helpful, as the code was
introduced in the initial commit of the template engine over 25 years
ago.