Do not clean caches of in-memory database in forks
If we register a function with `at_exit` before forking, this function will be
called at the termination of any child. In PR #2014, I registered such a function
to clean-up databases loaded with Ancient. As a result, the function is
called after each worker terminates. Alas, [Ancient.delete] is buggy and
cannot be used.

This bug is causing a memory leak in GeneWeb but it is acceptable
because the database is intented to stay in memory during the entire
execution of the server.

This PR removes the `at_exit` invokation in `load_database` and add
documentation to explain the caveat.