Create a separate library for logs
This commit creates a new sublibrary for GeneWeb named `Geneweb.logs`.
I believe a better design would involve using Logs library in `/lib` and
defining several sources.

The Syslog library should be a dependency of the gwd binary itself and
use Logs sources to output a log file. This modification is too
extensive for this PR and should be addressed in a separate one.