test/run_gw_test.sh in crl do not use number var
but use $cmd var to geneate file name.
This will avoid to pass a second parameter to crl, that is potential
source of error by developer.

I am placing in this same commit all my ref dir generated on Linux
Ubuntu, assuming we really want to keep on git.

Will be interesting to test this ref on a Mac.