Reorganize Sosa library
The Sosa library provides two implementations:
- One based on integer arrays.
- One based on Zarith library.

The user can select the implementation in `configure.ml`. The Zarith
implementation is the default, if Zarith is installed.

This PR reorganizes the Sosa library into a single, isolated directory `lib/sosa`.

Pros:
- The default remains unchanged. If Zarith is installed, Sosa based on Zarith
  is used. If not, Sosa based on integer arrays is used.
- Remove the `SOSA_PKG` variables from many `dune.in`. We are closer to
  remove the `dune.in` workaround!
- Use the stanza `select` instead of a trick with virtual libraries. In
  particular, `dune utop` works again (after merging #2165).

Cons:
- User cannot select manually the implementation of Sosa.

Important:
The `geneweb-rpc` package depends on Zarith because the `tls` package
requires it and we want to support the `tls` protocol for better security.
As a result:
```
opam switch create . --deps-only
```
will install Zarith (the output of opam pretends that Zarith is required
by the `geneweb` itself but I believe that it is a bug in opam).

If you want to compile Geneweb without Zarith, proceed as follows:
```shell
opam switch . --empty
opam install ./geneweb.opam --deps-only --with-test
make build-geneweb
```