Use `nread` to copy notes
In PR #2131, I wrote a new function to copy files on disk. This function
should be used to synchronize files in `outbase` but the notes can be in
memory only. A proper fix requires a complete refactoring of the notes,
which is to expensive.

This PR restores the previous code and fixes the issue #2168.