Ugly hack to remove polymorphic recursion in Templ
The mutually recursive function `interp_ast` in `Templ` contains a
polymorphic recursion. As type inference is indecidable for this kind of
recursion, the OCaml typechecker does its best and most of time needs
help with type annotations. Unfortunatly, it seems there is bug in these
annotations that has been solved between OCaml 4.08 and OCaml 4.14.

Currently, the template engine expects an environment of type `a env`
where `a` is polymorphic variable that represents the type of values
used by the template. Certain recursive calls of `interp_ast` expect an
environment with a different value type.

The hack consists in two parts:
1. Duplicate some functions in `Templ` in order to specialize them.
2. Pass a dummy value as argument of some functions because using a unit
   value in a recursive call introduces another polymorphic recursion.
   This value is not used by the code, so we could use `Obj.magic` but
   this solution is safer.

Notice that this hack is still better than the previous situation where
a global reference was used to close the recursion.