Do not catching the exception in `w_base`
As my comment explained, catching exceptions here makes the code harder
to debug. If the base is missing or something fails during the request
process, it is a fatal error if we don't catch it in the request
handler.

The previous code tried to output a 404 Not found error on the socket
but most of the time, it is too late as the HTTP header can be emitted
in the template engine.

For small pages, we could prepare the page in memory and, if everything works
as expected, we print the HTTP header and the page after. For large
page, users expect to get a partial result. These changes are too
extensive and complicated for the current code base.

This PR only removes the exception handler. As a result, the worker die
with such exceptions. I believe that this design is acceptable and we
can implement a better patch when the code base will be better.