Do not call Printexc.print_backtrace outside an exception handler
This function cannot be used outside exception handlers. We called it
our Logs module if the debug flag was set. As a result, we got an
irrelevant backtraces, even if there is no uncaught exceptions in
workers.

This commit also `enable_runtime_warnings` in debug mode. These warnings
are diisabled by default.