Allow definition of variables in simple templates
The PR #2232 introduces regressions in the template engine. The issue
comes from unbound variables in the evaluator function of the template
engine. The environment management in this engine is a complete mess.
Basically, we have (at least) three environments:
1. A builtin environment defined in the display modules (like `perso.ml`
   for instance). This environment is split into two parts: variables
   and functions and functions can be arbitrary OCaml code.
2. An environment for the template variables.
3. An environment for the template functions.

The environment for template variables is embedded into the builtin one
using `set_vother` and `get_vother` functions and these function MUST be
defined in each display modules...

In particular `Templ.output_builtin` didn't define them (it was dummy
implementation) because it expected environments that can only store
strings, which is not sufficient to implement them.

Changes:
1. The most important changes is to change the implementation of
   `Templ.output_builtin` and the type of its environment. It is a
   hotfix because better approaches are really extensive.
2. Improve sightly errors of the evaluator. We must catch exceptions
   throwing in the evaluator and print them, as some recursive calls
   destroy most of the backtrace and we cannot localize the origin of
   the error without this information.
3. Add a `cached` flag in the pipeline to disable caching for debugging
   purpose. Currently, I use the flag `Logs.debug_flag` to determine if
   we are in debugging mode. We should refactor this later after
   integrating Logs library and sources.