Do not capture closed input channel in `Database.with_database`
We do not execute again `Database.with_database` if the database is
loaded in main memory. In particular, all the channels open in this
function are closed after executing `Driver.load_database`. Alas,
we capture a closed channel in `index_of_string`. This commit
fixes this bug.

The inversed index for strings is now loaded in main memory when we use
Ancient.