Switch to logs and logs-syslog
This PR removes the `Geneweb_log` library, which was mainly a dummy
implementation of a logs library. `Logs` is the de facto standard
library for logging in modern OCaml.

As a good side effect, this PR moves or removes codes which are not
appropriate into GeneWeb library.

It also removes the syslog dependency. This library does not seem to be
maintained, whereas logs-syslog is still developed and used in MirageOS.

The `logs` library exposes 4 log level: `App`, `Info`, `Warn`, `Error` and
`Debug`. The default is as follows:
- `App` is sent to `stdout`,
- `Info`, `Warn`, `Error` and `Debug` are sent to the diagnostic
  channel, which is `stderr` by default.

By default, the `Debug` level is not enabled. You must use `-debug` to
get these messages.

The option `-log` of `gwd` can changed the diagnostic channel. You can
send it to `stdout`, `stderr`, a file or a udp syslog socket.

The following command:
```console
./distribution/gw/gwd -debug -log filename
```
launches a gwd server with a file diagnostic channel.

The following command:
```console
sudo setcap 'cap_net_bind_service=+ep' distribution/gw/gwd
./distrubution/gw/gwd -debug -log "<syslog>"
```
launches a gwd server with a Syslog diagnostic channel. Currently, it
only emit on `127.0.0.1:514`, which was the previous behavior. The 514
port is the default for Syslog logging service, but you need root
privileges to open a socket on this port.
You can test the feature with `socat`:
```console
sudo socat -u UDP-RECV:514 STDOUT
```